mod crypt extracted
