+++ rust/signer/src/lib.rs
@@ -30,6 +30,7 @@ extern crate jni;
 pub mod android;
 pub mod string;
 pub mod eth;
+pub mod crypt;
 pub mod bip39_mnemonic;
 
 use std::os::raw::c_void;
